数据删除改为假删除 计量设备删除时若有关联数据则禁止删除 修改缴费记录sql语句 收费页面费率不随用户类型改变 邮箱不为空时进行格式判断

+++ src/main/java/com/daye/sys/service/impl/TbJlsbServiceImpl.java
@@ -82,6 +82,11 @@ public class TbJlsbServiceImpl extends ServiceImpl<TbJlsbMapper, TbJlsb> impleme
     @Override
     @RequiredLog(operation = "删除计量设备")
     public JsonResult deleteJlsb(Long id) {
+        //判断设备上是否关联了数据
+        Integer num = tbjlsbmapper.findAssociatedData(id);
+        if(num>0){
+            return new JsonResult(new Throwable("存在相关数据，禁止删除！"));
+        }
         if (tbjlsbmapper.deleteById(id) == 1) {
             return new JsonResult("删除成功");
         }
